third_party/blink/web_tests/http/tests/devtools/sources/debugger/debugger-compile-and-run-expected.txt

Tests separate compilation and run.


Running: testSuccessfulCompileAndRun
Compiling script
Running script
Script result: 3

Running: testRunError
Compiling script
Running script
exceptionDetails:
   Uncaught ReferenceError: c is not defined
    at test.js:1:16
   line: 0, column: 15
   exceptionDetails stack trace:
       url: test.js
       function: 
       line: 0

Running: testCompileError
Compiling script
exceptionDetails:
   Uncaught SyntaxError: Unexpected token '}'
   line: 0, column: 0
   no stack trace attached to exceptionDetails
